added reviews added ffaker for tests

+++ db/migrate/20170611205153_create_restaurants.rb
@@ -3,7 +3,7 @@ class CreateRestaurants < ActiveRecord::Migration[5.1]
     create_table :restaurants do |t|
       t.string :name, null: false
       t.references :cuisine, null: false, foreign_key: true
-      t.integer :rating, null: false
+      t.float :rating, null: false
       t.boolean :accepts_10bis, null: false
       t.string :address, null: false
       t.integer :max_delivery_time_minutes, null: false
